+++ docs/changelog.md
@@ -3,6 +3,8 @@
 ## next
 - Added support for passing data between components via Context. See the [Usage
 docs](usage.md#passing-data-with-context) for more information. [PR #48](https://github.com/pelme/htpy/pull/48).
+- Added Django template backend. The Django template backend allows you to
+integrate htpy components directly with Django. [See the docs for more information](django.md#the-htpy-template-backend). [PR #37](https://github.com/pelme/htpy/pull/37).
 
 ## 24.8.1 - 2024-08-16
  - Added the `comment()` function to render HTML comments.